faas_translation/cloud_run_app/requirements.txt
flask==3.*
gunicorn==23.*
google-cloud-storage==2.*
